Move GroupApprovalTask to workflows submodule
